refactor: created sharedUi folder and moved common styles into it
