+++ src/components/button/types.ts
@@ -1,9 +1,6 @@
-import { Theme } from 'styles/Theme';
-
 export interface buttonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   variant: string;
-  theme: Theme;
 }
 export enum ButtonVariant {
   Black = 'black',
